LPS-160041 Add scroll to FragmentsSidebar
