+++ modules/apps/layout/layout-content-page-editor-web/src/main/resources/META-INF/resources/page_editor/plugins/fragments-widgets/components/TabsPanel.js
@@ -13,11 +13,14 @@
  */
 
 import ClayTabs from '@clayui/tabs';
+import {debounce} from 'frontend-js-web';
 import PropTypes from 'prop-types';
-import React from 'react';
+import React, {useEffect, useLayoutEffect, useRef} from 'react';
 
 import {FRAGMENTS_DISPLAY_STYLES} from '../../../app/config/constants/fragmentsDisplayStyles';
+import {config} from '../../../app/config/index';
 import {useId} from '../../../app/utils/useId';
+import {useSessionState} from '../../../core/hooks/useSessionState';
 import {COLLECTION_IDS} from './FragmentsSidebar';
 import TabCollection from './TabCollection';
 
@@ -33,10 +36,57 @@ export default function TabsPanel({
 
 	const getTabId = (tabId) => `${tabIdNamespace}tab${tabId}`;
 	const getTabPanelId = (tabId) => `${tabIdNamespace}tabPanel${tabId}`;
+	const wrapperElementRef = useRef(null);
+
+	const [scrollPosition, setScrollPosition] = useSessionState(
+		`${config.portletNamespace}_fragments-sidebar_${activeTabId}_scroll-position`,
+		0,
+		{
+			persistEnabled: Liferay.FeatureFlags['LPS-153452'],
+		}
+	);
+
+	const scrollPositionRef = useRef(scrollPosition);
+	scrollPositionRef.current = scrollPosition;
+
+	useLayoutEffect(() => {
+		const wrapperElement = wrapperElementRef.current;
+		const initialScrollPosition = scrollPositionRef.current;
+
+		if (!wrapperElement || !initialScrollPosition) {
+			return;
+		}
+
+		wrapperElement.scrollBy({
+			behavior: 'auto',
+			left: 0,
+			top: initialScrollPosition,
+		});
+	}, []);
+
+	useEffect(() => {
+		const wrapperElement = wrapperElementRef.current;
+
+		if (!wrapperElement) {
+			return;
+		}
+
+		const handleScroll = debounce(() => {
+			setScrollPosition(wrapperElement.scrollTop);
+		}, 300);
+
+		wrapperElement.addEventListener('scroll', handleScroll, {
+			passive: true,
+		});
+
+		return () => {
+			wrapperElement.removeEventListener('scroll', handleScroll);
+		};
+	}, [setScrollPosition]);
 
 	return (
 		<>
-			<ClayTabs className="mb-4 page-editor__sidebar__fragments-widgets-panel__tabs">
+			<ClayTabs className="flex-shrink-0 mb-4 mx-3 page-editor__sidebar__fragments-widgets-panel__tabs">
 				{tabs.map((tab, index) => (
 					<ClayTabs.Item
 						active={tab.id === activeTabId}
@@ -45,44 +95,55 @@ export default function TabsPanel({
 							'id': getTabId(index),
 						}}
 						key={index}
-						onClick={() => setActiveTabId(tab.id)}
+						onClick={() => {
+							setActiveTabId(tab.id);
+							setScrollPosition(0);
+
+							if (wrapperElementRef.current) {
+								wrapperElementRef.current.scrollTop = 0;
+							}
+						}}
 					>
 						{tab.label}
 					</ClayTabs.Item>
 				))}
 			</ClayTabs>
 
-			<ClayTabs.Content
-				activeIndex={tabs.findIndex((tab) => tab.id === activeTabId)}
-				className="page-editor__sidebar__fragments-widgets-panel__tab-content"
-				fade
-			>
-				{tabs.map((tab, index) => (
-					<ClayTabs.TabPane
-						aria-labelledby={getTabId(index)}
-						id={getTabPanelId(index)}
-						key={index}
-					>
-						<ul className="list-unstyled">
-							{tab.collections.map((collection, index) => (
-								<TabCollection
-									collection={collection}
-									displayStyle={
-										tab.id === COLLECTION_IDS.widgets
-											? FRAGMENTS_DISPLAY_STYLES.LIST
-											: displayStyle
-									}
-									initialOpen={
-										index <
-										INITIAL_EXPANDED_ITEM_COLLECTIONS
-									}
-									key={index}
-								/>
-							))}
-						</ul>
-					</ClayTabs.TabPane>
-				))}
-			</ClayTabs.Content>
+			<div className="overflow-auto px-3" ref={wrapperElementRef}>
+				<ClayTabs.Content
+					activeIndex={tabs.findIndex(
+						(tab) => tab.id === activeTabId
+					)}
+					className="page-editor__sidebar__fragments-widgets-panel__tab-content"
+					fade
+				>
+					{tabs.map((tab, index) => (
+						<ClayTabs.TabPane
+							aria-labelledby={getTabId(index)}
+							id={getTabPanelId(index)}
+							key={index}
+						>
+							<ul className="list-unstyled">
+								{tab.collections.map((collection, index) => (
+									<TabCollection
+										collection={collection}
+										displayStyle={
+											tab.id === COLLECTION_IDS.widgets
+												? FRAGMENTS_DISPLAY_STYLES.LIST
+												: displayStyle
+										}
+										initialOpen={
+											index <
+											INITIAL_EXPANDED_ITEM_COLLECTIONS
+										}
+										key={index}
+									/>
+								))}
+							</ul>
+						</ClayTabs.TabPane>
+					))}
+				</ClayTabs.Content>
+			</div>
 		</>
 	);
 }
